# infrastructure/handlers/driverHandler.go
package handlers

import (
	"log"
	"net/http"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/gocql/gocql"
	"github.com/vinaocruz/go-scylla-example/database"
	"github.com/vinaocruz/go-scylla-example/domain/entities"
	domain "github.com/vinaocruz/go-scylla-example/domain/repositories"
	infra "github.com/vinaocruz/go-scylla-example/infrastructure/repositories"
)

type DriverHandler struct {
	repository domain.DriverRepository
}

func NewDriverHandler(db *database.Database, router *gin.RouterGroup) {
	handler := &DriverHandler{
		repository: infra.NewScyllaDriverRepository(db),
	}

	router.GET("/", handler.ListDrivers)
	router.POST("/", handler.CreateDriver)
	router.GET("/:cnh", handler.GetDriver)
	router.PUT("/:cnh", handler.UpdateDriver)
	router.DELETE("/:cnh", handler.DeleteDriver)
}

// @Summary lista motoristas
// @Description Busca a lista de todos motoristas e seus veiculos
// @Tags Motoristas
// @Accept json
// @Produce json
// @Success 200 {array} entities.Driver
// @failure 500 {string} string "failed to list all"
// @Router /drivers [get]
func (handler *DriverHandler) ListDrivers(c *gin.Context) {
	listAll, err := handler.repository.ListAll()

	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{
			"error": "failed to list all",
		})
		return
	}

	c.JSON(http.StatusOK, listAll)
}

// @Summary cria novo motorista e veiculo
// @Description Este endpoint permite adicionar um novo motorista e seu veiculo
// @Tags Motoristas
// @Accept json
// @Produce json
// @Param data body string true "Numero de CNH"
// @Param name body string true "Nome do motorista"
// @Param license_plate body string true "Número da plata"
// @Param model body string true "Modelo do veiculo"
// @Success 200 {object} entities.Driver
// @failure 400 {string} string "failed to bind json"
// @failure 500 {string} string "failed to store"
// @Router /drivers [post]
func (handler *DriverHandler) CreateDriver(c *gin.Context) {
	var driver entities.Driver

	if err := c.ShouldBindJSON(&driver); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{
			"error": "failed to bind json",
		})
		return
	}

	driver.ID = gocql.TimeUUID()
	driver.CreatedAt = time.Now()

	err := handler.repository.Store(driver)
	if err != nil {
		log.Fatal(err)
		c.JSON(http.StatusInternalServerError, gin.H{
			"error": "failed to store",
		})
		return
	}

	c.JSON(http.StatusOK, driver)
}

// @Summary buscar um motorista
// @Description Este endpoint permite buscar dados e veiculos de um motorista identificado
// @Tags Motoristas
// @Accept json
// @Produce json
// @Success 200 {array} entities.Driver
// @failure 404 {string} string
// @Router /drivers/{cnh} [get]
// @Param cnh path string true "Número de CNH"
func (handler *DriverHandler) GetDriver(c *gin.Context) {
	drivers, err := handler.repository.Load(c.Param("cnh"))
	if err != nil {
		c.JSON(http.StatusNotFound, gin.H{
			"error": err.Error(),
		})
		return
	}

	c.JSON(http.StatusOK, drivers)
}

// @Summary edita motorista
// @Description Este endpoint permite alterar dados do motorista
// @Tags Motoristas
// @Accept json
// @Produce json
// @Success 200 {array} entities.Driver
// @failure 404 {string} string
// @failure 400 {string} string "failed to bind json"
// @failure 500 {string} string "failed to update"
// @Router /drivers/{cnh} [put]
// @Param cnh path string true "Número de CNH"
// @Param name body string true "Nome do motorista"
func (handler *DriverHandler) UpdateDriver(c *gin.Context) {
	cnh := c.Param("cnh")

	drivers, err := handler.repository.Load(cnh)
	if err != nil {
		c.JSON(http.StatusNotFound, gin.H{
			"error": err.Error(),
		})
		return
	}

	var people entities.People

	if err := c.ShouldBindJSON(&people); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{
			"error": "failed to bind json",
		})
		return
	}

	for i := range drivers {
		drivers[i].Name = people.Name
		err = handler.repository.Store(drivers[i])

		if err != nil {
			c.JSON(http.StatusInternalServerError, gin.H{
				"error": "failed to update",
			})
			return
		}
	}

	c.JSON(http.StatusOK, drivers)
}

// @Summary remove motorista
// @Description Este endpoint remove dados do motorista e seus veiculos
// @Tags Motoristas
// @Accept json
// @Produce json
// @Success 204
// @failure 500 {string} string "failed to delete"
// @Router /drivers/{cnh} [delete]
// @Param cnh path string true "Número de CNH"
func (handler *DriverHandler) DeleteDriver(c *gin.Context) {
	err := handler.repository.Delete(c.Param("cnh"))
	if err != nil {
		c.JSON(http.StatusInternalServerError, gin.H{
			"error": "failed to delete",
		})
		return
	}

	c.JSON(http.StatusNoContent, gin.H{})
}